converting ipynb to pytest

+++ chaospy/quadrature/discrete.py
@@ -60,7 +60,7 @@ def discrete_simple(order, lower=-2, upper=2):
 
     Use :func:`chaospy.quadrature.discrete` instead.
     """
-    order = min(order, round(upper-lower)-1)
+    order = int(min(order, round(upper-lower)-1))
     abscissas = numpy.linspace(lower, upper, 2*order+3)[1::2].round()
     weights = numpy.full(order+1, (upper-lower)/(order+1.))
     return abscissas, weights
